refactor: move symbols into a dedicated constants file
