+++ src/constants.js
@@ -1,0 +1,5 @@
+/* @flow */
+
+export const DONE: Symbol = Symbol('done')
+export const SKIP: Symbol = Symbol('skip')
+
