+++ src/from.js
@@ -1,7 +1,8 @@
 /* @flow */
 
+import { DONE, SKIP } from './constants'
 import Base from './base'
-import { DONE, SKIP, apply, done } from './adapter'
+import { apply, done } from './adapter'
 
 export default class From<T> extends Base<T> {
   iterator: Iterator<T>
